+++ src/g_game.h
@@ -113,18 +113,31 @@ extern int  bodyquesize, default_bodyquesize; // killough 2/8/98, 10/98
 extern int pars[][10];  // hardcoded array size
 extern int cpars[];     // hardcoded array size
 
-// [Nugget] ------------------------------------------------------------------
+// [Nugget] ==================================================================
 
 void G_SetAutosaveCountdown(int value); // Autosave
 
-// Rewind
+// Rewind --------------------------------------------------------------------
+
 void G_SetRewindCountdown(int value);
 void G_EnableRewind(void);
 void G_Rewind(void);
 void G_ClearExcessKeyFrames(void);
 boolean G_KeyFrameRW(void);
 
-// Skill
+// Slow Motion ---------------------------------------------------------------
+
+#define SLOWMO_FACTOR_TARGET 0.33f
+#define SLOWMO_FACTOR_NORMAL 1.0f
+
+boolean G_GetSlowMotion(void);
+void G_SetSlowMotion(const boolean value);
+void G_ResetSlowMotion(void);
+float G_GetSlowMotionFactor(void);
+void G_SetTimeScale(int scale);
+
+// Skill ---------------------------------------------------------------------
+
 void G_SetSkillParms(const skill_t skill);
 void G_SetUserCustomSkill(void);
 void G_RestartWithLoadout(const boolean current);