+++ src/p_map.h
@@ -86,11 +86,23 @@ extern struct msecnode_s *sector_list;                             // phares 3/1
 extern fixed_t tmbbox[4];         // phares 3/20/98
 extern struct line_s *blockline;   // killough 8/11/98
 
-// [Nugget] ------------------------------------------------------------------
+// [Nugget] ==================================================================
 
 fixed_t P_PitchToSlope(const fixed_t pitch);
 
-void P_PositionChasecam(fixed_t z, fixed_t dist, fixed_t slope);
+void P_PositionChasecam(fixed_t z, fixed_t dist, fixed_t slope); // Chasecam
+
+extern boolean boomshot; // Explosive hitscan cheat
+
+// Hitscan trails ------------------------------------------------------------
+
+int  P_GetShowHitscanTrails(void);
+int  P_CycleShowHitscanTrails(void);
+void P_SpawnHitscanTrail(fixed_t x, fixed_t y, fixed_t z,
+                         angle_t angle, fixed_t slope,
+                         fixed_t range, fixed_t distance);
+
+// Over/Under ----------------------------------------------------------------
 
 typedef enum {
   OU_UNDER = -1,
@@ -98,11 +110,8 @@ typedef enum {
   OU_OVER,
 } overunder_t;
 
-// Over/Under
 overunder_t P_CheckOverUnderMobj(struct mobj_s *thing); // [DSDA]
 boolean     P_SkullSlam(struct mobj_s **skull, struct mobj_s *hitthing);
-
-extern boolean boomshot; // Explosive hitscan cheat
 
 #endif // __P_MAP__