+++ src/s_sound.c
@@ -38,6 +38,7 @@
 #include "z_zone.h"
 
 // [Nugget]
+#include "g_game.h"
 #include "r_main.h"
 #include "r_state.h"
 
@@ -232,6 +233,10 @@ void S_StartSoundPitch(const mobj_t *origin, int sfx_id, const pitchrange_t pitc
 
     // Initialize sound parameters
     pitch = NORM_PITCH;
+
+    // [Nugget] Slow Motion
+    if (!menuactive && G_GetSlowMotionFactor() != SLOWMO_FACTOR_NORMAL)
+    { pitch *= G_GetSlowMotionFactor(); }
 
     // haleyjd: modified so that priority value is always used
     // haleyjd: also modified to get and store proper singularity value